TST: Move some window/moments tests to test_ewm
